moving all migration logic to the new package
