+++ pkg/pgmodel/migrate/migrate.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/jackc/pgx/v4"
 	"github.com/timescale/promscale/pkg/log"
-	"github.com/timescale/promscale/pkg/pgmodel"
 	"github.com/timescale/promscale/pkg/pgmodel/common/extension"
 	"github.com/timescale/promscale/pkg/util"
 )
@@ -14,7 +13,7 @@ var (
 	MigrationLockError = fmt.Errorf("could not acquire migration lock. Ensure there are no other connectors running and try again")
 )
 
-func SetupDBState(conn *pgx.Conn, appVersion pgmodel.VersionInfo, leaseLock *util.PgAdvisoryLock, extOptions extension.ExtensionMigrateOptions) error {
+func SetupDBState(conn *pgx.Conn, appVersion VersionInfo, leaseLock *util.PgAdvisoryLock, extOptions extension.ExtensionMigrateOptions) error {
 	// At startup migrators attempt to grab the schema-version lock. If this
 	// fails that means some other connector is running. All is not lost: some
 	// other connector may have migrated the DB to the correct version. We warn,
@@ -39,7 +38,7 @@ func SetupDBState(conn *pgx.Conn, appVersion pgmodel.VersionInfo, leaseLock *uti
 		log.Warn("msg", "skipping migration lock")
 	}
 
-	err := pgmodel.Migrate(conn, appVersion, extOptions)
+	err := Migrate(conn, appVersion, extOptions)
 	if err != nil {
 		return fmt.Errorf("error while trying to migrate DB: %w", err)
 	}
